requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --allow-unsafe --output-file=./requirements.txt ./setup/requirements.prod.in
#
asgiref==3.2.7            # via django
django==3.0.6             # via -r ./setup/requirements.base.in, djangorestframework
djangorestframework==3.11.0  # via -r ./setup/requirements.base.in
gunicorn==20.0.4          # via -r ./setup/requirements.base.in
psycopg2-binary==2.8.5    # via -r ./setup/requirements.prod.in
pyjwt==1.7.1              # via -r ./setup/requirements.base.in
python-dotenv==0.13.0     # via -r ./setup/requirements.base.in
pytz==2020.1              # via django
sqlparse==0.3.1           # via django

# The following packages are considered to be unsafe in a requirements file:
setuptools==46.3.1        # via gunicorn
